+++ config/splat.us.dra.yaml
@@ -106,7 +106,7 @@ segments:
       - [0xCF4C, .data, 6E42C]
       - [0xCF6C, .data, 70394]
       - [0xCF74, .data, d_CF74]
-      - [0xD0C4, .data, 7A4D0]
+      - [0xD0C4, .data, 7879C]
       - [0xD1D4, assets, factory_blueprint, g_FactoryBlueprints]
       - [0xD4B8, .data, 7E4BC]
       - [0xDBD4, .data, d_DBD4]
@@ -210,10 +210,10 @@ segments:
       # - [0x41A38, .rodata, 75F54] # func_80117DEC
       - [0x41A44, .rodata, 7879C] # EntityNumberMovesToHpMeter
       # - [0x41A74, .rodata, 7879C] # EntityGuardText
-      - [0x41AA4, .rodata, 7A4D0] # func_8011A4D0
-      # - [0x41AB4, .rodata, 7A4D0] # EntityEntFactory
-      # - [0x41AF4, .rodata, 7A4D0] # EntityPlayerBlinkWhite
-      # - [0x41B94, .rodata, 7A4D0] # EntityMpReplenished
+      # - [0x41AA4, .rodata, 7879C] # func_8011A4D0
+      # - [0x41AB4, .rodata, 7879C] # EntityEntFactory
+      # - [0x41AF4, .rodata, 7879C] # EntityPlayerBlinkWhite
+      # - [0x41B94, .rodata, 7879C] # EntityMpReplenished
       - [0x41CB4, .rodata, 7E4BC] # func_8011E4BC
       # - [0x41D84, .rodata, 7E4BC] # EntityPlayerPinkEffect
       # - [0x41DA4, .rodata, 7E4BC] # EntityPlayerDissolves
@@ -267,7 +267,6 @@ segments:
       - [0x72BB0, c, 72BB0]
       - [0x75F54, c, 75F54]
       - [0x7879C, c, 7879C]
-      - [0x7A4D0, c, 7A4D0]
       - [0x7E4BC, c, 7E4BC]
       - [0x843B0, c, 843B0]
       - [0x84B88, c, 84B88]
